requirements.txt
openai
rss_parser
python-slugify
bs4
python-dotenv